requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
frozenlist==1.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
pyee==11.0.0
python-dotenv==1.0.0
slack-sdk==3.21.3
slackeventsapi==3.0.1
typing_extensions==4.7.1
Werkzeug==2.3.7
yarl==1.9.2
